+++ pkg/mode/job/constants.go
@@ -14,10 +14,12 @@
 
 package job
 
+import "talend/vault-sidecar-injector/pkg/config"
+
 const (
 	//--- Job handling - Temporary mechanism until KEP https://github.com/kubernetes/enhancements/blob/master/keps/sig-apps/sidecarcontainers.md is implemented (and we migrate on appropriate version of k8s)
-	jobMonitoringContainerName = "tvsi-job-babysitter" // Name of our specific sidecar container to inject in submitted jobs
-	jobListenerContainerName   = "tvsi-vault-agent"    // Name of the container listening for signal from job monitoring container
+	jobMonitoringContainerName = config.JobMonitoringContainerName // Name of our specific sidecar container to inject in submitted jobs
+	jobListenerContainerName   = config.VaultAgentContainerName    // Name of the container listening for signal from job monitoring container
 
 	//--- Job handling env vars
 	jobContainerNameEnv = "VSI_JOB_CNT_NAME" // Env var for name of the app job's container